+++ src/Dap.Build/RELEASE_NOTES.md
@@ -1,3 +1,8 @@
+## 0.6.3
+* Generate .sha512 file when inject
+* Save original version before inject
+* Create Recover target to remove the injected version
+
 ## 0.6.2
 * Create targets for each projects